pyproject.toml
[project]
name = "bankoin"
version = "0.1.0"
description = "Desafio de Projeto: API Bancária Assíncrona com FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]
